+++ Excellia/Excellia/WorldManager.h
@@ -96,7 +96,7 @@ private:
 	std::vector<int> m_dirt_heights;	/* This is how much dirt should be on top */
 	std::vector<int> m_heights;
 	std::vector<int> m_trees;
-	std::vector<sf::Vector2f> m_caves;
+	std::vector<sf::Vector2i> m_caves;
 	sf::Image m_image;
 	sf::Texture m_texture;
 	sf::Sprite m_sprite;
